pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ishifuku-gold-scraper"
version = "1.0.0"
description = "石福金属興業から金価格を取得するスクレイピングツール"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.0",
    "pytest-xdist>=3.5.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "types-requests>=2.32.0",
    "types-beautifulsoup4>=4.12.0",
    "boto3-stubs[s3]>=1.34.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ishifuku-gold-scraper"
Repository = "https://github.com/yourusername/ishifuku-gold-scraper.git"
Issues = "https://github.com/yourusername/ishifuku-gold-scraper/issues"

[project.scripts]
ishifuku-scraper = "scrape_ishifuku:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=scrape_ishifuku",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | tests
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true      # 段階的に有効化
disallow_incomplete_defs = true   # 段階的に有効化
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
explicit_package_bases = true
namespace_packages = true

# モジュール別設定
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "main_lambda",
    "main_local",
    "optimized_core"
]
# テストとメインモジュールは段階的に厳格化
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
warn_return_any = false

[[tool.mypy.overrides]]
module = ["pytest", "boto3", "botocore.*", "selenium.*", "webdriver_manager.*"]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "tests",
]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["tests/*"]
